[feat/#109] - 참가자들 정보에 대한 useParticipantsQuery 커스텀 훅 생성하여 코드 간결화
