--- src/components/chat/chatHeader/ChatParticipants.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const ChatParticipants = () => {
-  return <div>ChatParticipants</div>;
-};
-
-export default ChatParticipants;
