+++ src/hooks/useQueries/useMeetingQuery.ts
@@ -1,5 +1,5 @@
-import { fetchMyRoom, fetchRecruitingRoom } from '(@/services/queryFns)';
-import { MY_ROOM, RECRUTING_ROOMDATA } from '(@/services/queryKeys)';
+import { fetchMyRoom, fetchRecruitingRoom } from '(@/query/meetingRoom/meetingQueryFns)';
+import { MY_ROOM, RECRUTING_ROOMDATA } from '(@/query/meetingRoom/meetingQueryKeys)';
 import { useSuspenseQueries } from '@tanstack/react-query';
 
 export const useRecruitingMyroomQuery = (user_id: string) => {
@@ -15,6 +15,5 @@ export const useRecruitingMyroomQuery = (user_id: string) => {
       }
     ]
   });
-
-  return results;
+  return results.map((re) => re.data);
 };
