+++ src/query/chat/chatQueryKeys.ts
@@ -1,0 +1,3 @@
+export const ROOMDATA_QUERY_KEY = ['room'];
+
+export const PARTICIPANTS_QUERY_KEY = ['paricipants'];
