+++ src/query/meetingRoom/meetingQueryKeys.ts
@@ -1,5 +1,3 @@
-export const ROOMDATA_QUERY_KEY = ['room'];
-
 export const RECRUTING_ROOMDATA = ['rercuiting'];
 
 export const MY_ROOM = ['myRoom'];
